rename qsvm-variational -> vqc
